pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vcenter-exporter"
version = "0.1.0"
description = "Prometheus exporter for vCenter (clusters, hosts, storage, VMs) via vSphere Automation API"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Influence Digital" }]
keywords = ["vcenter", "vsphere", "prometheus", "exporter", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "requests>=2.28.0",
    "prometheus_client>=0.19.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]

[project.scripts]
vcenter-exporter = "vcenter_exporter.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["vcenter_exporter*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
